backend/requirements.txt
flask
flask-cors
google-auth
mysql-connector-python